+++ src/main/java/com/archisemtle/semtlewebserverspring/infrastructure/ApplicationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApplicationRepository extends JpaRepository<Application, Long> {
+public interface ApplicationRepository extends JpaRepository<Application, Integer> {
 
-    Page<Application> findByApplicantId(Long applicantId, Pageable pageable);
+    Page<Application> findByApplicantId(Integer applicantId, Pageable pageable);
 }
